Chore: separated merger implementation classes in different files
