+++ lib/flutter_quill_delta_easy_parser.dart
@@ -7,4 +7,7 @@ export 'package:flutter_quill_delta_easy_parser/blocks/text_fragment.dart';
 export 'package:flutter_quill_delta_easy_parser/easy_parser/document_parser.dart';
 export 'package:flutter_quill_delta_easy_parser/easy_parser/easy_parser.dart';
 export 'package:flutter_quill_delta_easy_parser/easy_parser/merger_builder.dart';
+export 'package:flutter_quill_delta_easy_parser/easy_parser/block_merger_builder.dart';
+export 'package:flutter_quill_delta_easy_parser/easy_parser/common_merger_builder.dart';
+export 'package:flutter_quill_delta_easy_parser/easy_parser/no_merger_builder.dart';
 export 'package:flutter_quill_delta_easy_parser/enums/enums.dart';
